service layers completed

+++ src/main/java/com/travelbooking/travelbooking/dto/FlightRequestDto.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -17,5 +16,5 @@ public class FlightRequestDto {
     private String fromCity;
     private String toCity;
     private LocalDate flightDate;
-    private BigDecimal price;
+    private Double price;
 }
